requirements.txt
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.28.1
python-dotenv==0.21.0
google-generativeai==0.8.5
openai==1.82.0